+++ src/pages/NewAccount/Vanity.vue
@@ -397,6 +397,7 @@ const useVanityStartClick = () => {
           </Button>
           <QRCodeVue3
             v-if="state.showQR"
+            imgclass="qr-code"
             :width="500"
             :height="500"
             :value="state.w"