+++ src/pages/ReceivePayment.vue
@@ -121,6 +121,7 @@
           <label class="col-12 mb-2 md:col-2 md:mb-0"></label>
           <div class="col-12 md:col-10">
             <QRCodeVue3
+              imgclass="qr-code"
               :width="400"
               :height="400"
               :value="qrcode"